+++ src/ngx_http_haskell_load.c
@@ -214,6 +214,16 @@ ngx_http_haskell_load(ngx_cycle_t *cycle)
         goto dlclose_and_exit;
     }
 
+    mcf->set_cached_pid = (void (*)(HsInt32)) dlsym(mcf->dl_handle,
+                                            "ngxExportSetCachedPid");
+    dl_error = dlerror();
+    if (dl_error != NULL) {
+        ngx_log_error(NGX_LOG_EMERG, cycle->log, 0,
+                      "failed to load function "
+                      "\"ngxExportSetCachedPid\": %s", dl_error);
+        goto dlclose_and_exit;
+    }
+
     install_signal_handler = (void (*)(void)) dlsym(mcf->dl_handle,
                                             "ngxExportInstallSignalHandler");
     dl_error = dlerror();