+++ src/ngx_http_haskell_module.c
@@ -640,6 +640,7 @@ ngx_http_haskell_init_worker(ngx_cycle_t *cycle)
 
     mcf->set_upstream_main_conf_ptr(umcf);
     mcf->set_cached_time_ptr((volatile void **) &ngx_cached_time);
+    mcf->set_cached_pid(ngx_pid);
 
     service_hooks = mcf->service_hooks.elts;
     for (i = 0; i < mcf->service_hooks.nelts; i++) {